+++ src/http.h
@@ -23,13 +23,13 @@
 namespace sasl_xoauth2 {
 
 using HttpIntercept = std::function<int(
-    const std::string &url, const std::string &data, int *response_code,
+    const std::string &url, const std::string &data, long *response_code,
     std::string *response, std::string *error)>;
 
 void SetHttpInterceptForTesting(HttpIntercept intercept);
 
 int HttpPost(const std::string &url, const std::string &data,
-             int *response_code, std::string *response, std::string *error);
+             long *response_code, std::string *response, std::string *error);
 
 }  // namespace sasl_xoauth2
